refactor(graphs): split capability implementations by outcome
