+++ src/jacobian/graphs/__init__.py
@@ -1,6 +1,5 @@
 """Graph construction, property, coloring, composition, and shrinking capabilities."""
 
-from jacobian.graphs.capabilities import GraphInstallation, install_graph_capabilities
 from jacobian.graphs.coloring import (
     GraphColoringInstallation,
     install_graph_coloring_capabilities,
@@ -9,6 +8,7 @@ from jacobian.graphs.composition import (
     GraphCompositionInstallation,
     install_graph_composition_capabilities,
 )
+from jacobian.graphs.installation import GraphInstallation, install_graph_capabilities
 from jacobian.graphs.isomorphism import (
     GraphIsomorphismInstallation,
     install_graph_isomorphism,
